% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.base.R
\name{plot.base}
\alias{plot.base}
\title{Plot the estimated baseline function}
\usage{
\method{plot}{base}(base, main = NULL, xlab = "time", ylab = "baseline")
}
\arguments{
\item{base}{the output from recbase().}

\item{main}{title of the plot.}

\item{xlab}{label of the x axis.}

\item{ylab}{label of the y axis.}
}
\description{
Plot the baseline function estimated by recbase().
}
