% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.recdata.R
\name{plot.recdata}
\alias{plot.recdata}
\title{Plot the recurrent event data}
\usage{
\method{plot}{recdata}(recdata, time, ids = NULL, n.id = 5, main = NULL,
  xlab = "time", ylab = "id")
}
\arguments{
\item{recdata}{a rectime data object generated from recdata().}

\item{time}{the column name for the time variable.}

\item{ids}{the vector of subject id that the user wants to plot. The default is NULL. If
ids=NULL, the first n.id subjects will be plotted.}

\item{n.id}{the number of subjects that the user wants to plot. The default is 5.}

\item{main}{the title of the plot. The default is NULL.}

\item{xlab}{the label of the x axis. The default is "time".}

\item{ylab}{the label of the y axis. The default is "id".}
}
\description{
Plot the recurrent events and regular visits for several subjects for illustration.
}
\examples{
data(simdata)
recd <- recdata(id="id", type="type", time="time", data=simdata)
plot.recdata(recd, time="time", n.id=5, xlab="time", ylab="id")
}
