% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recdata.R
\name{recdata}
\alias{recdata}
\title{Create the rectime data object}
\usage{
recdata(id, type, time, data)
}
\arguments{
\item{id}{the column name for the subject id in the dataset.}

\item{type}{the column name for the variable which indicates the type of each observation:
1: event visit; 2: regular visit; 0: censoring.}

\item{time}{the column name for the time variable.}

\item{data}{the dataset. Must include: a column for the subject id, a column for the
observation type, a column for the time variable and several columns for the covariates. Note
that we assume that time variable starts from 0 and it needs to be 'numeric'. The covariates
need to be 'numeric' as well.}
}
\value{
A list of two datasets: d.event and d.regular.
}
\description{
Create the rectime data object used as input of other functions.
}
\details{
The rectime data object is a list of two datasets: d.event is for the event observations;
d.regular is for the observations at regular visits. The observations with a missing time variable
will be deleted: for d.event, the whole subject will be deleted if the time variable is
missing in one observation; for d.regular, only the observation with missing time variable
will be deleted. In addition, in d.event and d.regular, the column names for subject id,
type of observation and censoring time are fixed as: "id", "type", "cent".
}
\examples{
data(simdata)
recd <- recdata(id="id", type="type", time="time", data=simdata)
}
